requirements.txt
altair==5.5.0
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
imbalanced-learn==0.12.4
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib==3.9.4
narwhals==2.18.1
numba==0.60.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xgboost==2.1.4
zipp==3.23.0
